voice-assistant/requirements.txt
# Audio I/O
PyAudio==0.2.14
webrtcvad==2.0.10
pydub==0.25.1

# Google Cloud APIs
google-cloud-speech==2.26.0
google-cloud-texttospeech==2.16.3
google-auth==2.34.0

# Anthropic Claude
anthropic==0.34.2

# Configuration
python-dotenv==1.0.1

# GPIO (optional - only for button trigger mode)
gpiozero==2.0.1

# Standard library supplements
requests==2.32.3
